:broom: chore(api): refactor complementaryCertification subdomain
