+++ api/tests/certification/configuration/integration/infrastucture/repositories/complementary-certification-badge-repository_test.js
@@ -3,7 +3,7 @@ import lodash from 'lodash';
 import { databaseBuilder, domainBuilder, expect, knex, sinon } from '../../../../../test-helper.js';
 
 const { omit } = lodash;
-import * as complementaryCertificationBadgeRepository from '../../../../../../src/certification/complementary-certification/infrastructure/repositories/complementary-certification-badge-repository.js';
+import * as complementaryCertificationBadgeRepository from '../../../../../../src/certification/configuration/infrastructure/repositories/complementary-certification-badge-repository.js';
 import { DomainTransaction } from '../../../../../../src/shared/domain/DomainTransaction.js';
 
 describe('Integration | Infrastructure | Repository | Certification | Complementary-certification | complementary-certification-badge-repository', function () {
