+++ api/tests/certification/configuration/integration/infrastucture/repositories/organization-repository_test.js
@@ -1,4 +1,4 @@
-import * as organizationRepository from '../../../../../../src/certification/complementary-certification/infrastructure/repositories/organization-repository.js';
+import * as organizationRepository from '../../../../../../src/certification/configuration/infrastructure/repositories/organization-repository.js';
 import { databaseBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Integration | Repository | Certification | Complementary-certification | Organization', function () {
